baler4 - Restructure baler for more flexible parsing
- Added a parser front end that is independent of the transport
- Added additional types for tokens for more rich patterns
- Added a generic storage interface so that multiple storage
  architectures could be explored without tearing up baler
- Added a python interface to the storage API so that web
  based browsing of baler data is easily integrated with existing
  web infrastructures such as OpenStack

+++ src/plugins/bout_sos_msg.c
@@ -117,7 +117,7 @@ int bout_sos_msg_process_output(struct boutplugin *this,
 	msg->data.uint32_[BSOS_MSG_COMP_ID] = odata->comp_id;
 	msg->data.uint32_[BSOS_MSG_PTN_ID] = odata->msg->ptn_id;
 	for (arg = 0; arg < odata->msg->argc; arg++)
-		msg->data.uint32_[arg + BSOS_MSG_ARGV_0] = odata->msg->argv[arg];
+		msg->data.uint32_[arg + BSOS_MSG_ARGV_0] = (uint32_t)odata->msg->argv[arg];
 
 	/* creat key */
 	ptc_k.comp_id = msg->data.uint32_[BSOS_MSG_COMP_ID];
